+++ java-generator/it/src/it/escape-characters/src/test/java/io/fabric8/it/dummy/TestEscapeCharacters.java
@@ -18,6 +18,8 @@ package io.fabric8.it.dummy;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.example.v1.Dummy;
 import com.example.v1.DummySpec;
+import com.example.v1.dummyspec.Package;
+import com.example.v1.dummyspec.package_.Foo;
 import io.fabric8.kubernetes.api.model.ObjectMeta;
 import io.fabric8.kubernetes.client.utils.Serialization;
 import org.junit.jupiter.api.Test;
@@ -48,6 +50,7 @@ class TestEscapeCharacters {
     assertEquals("3", spec.getThree_quote());
     assertEquals("4", spec.getFour_doublequote());
     assertEquals("5", spec.getFive_slash());
+    assertEquals("6", spec.get_package().getFoo().getBar());
   }
 
   @Test
@@ -63,6 +66,11 @@ class TestEscapeCharacters {
     spec.setThree_quote("3");
     spec.setFour_doublequote("4");
     spec.setFive_slash("5");
+    Foo foo = new Foo();
+    foo.setBar("6");
+    Package pack = new Package();
+    pack.setFoo(foo);
+    spec.set_package(pack);
     sample.setSpec(spec);
     ObjectMeta om = new ObjectMeta();
     om.setName("sample");